ConstProp: Merge ConstantPropagationAnalysis into ConstPropTransformer
